+++ backend/LoyaltyApi/Services/ICreditPointsTransactionService.cs
@@ -1,7 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using System.Numerics;
 using LoyaltyApi.Models;
 using LoyaltyApi.RequestModels;
 
@@ -11,7 +8,7 @@ namespace LoyaltyApi.Services
     {
         public Task<CreditPointsTransaction?> GetTransactionByIdAsync(int transactionId);
 
-        public Task<CreditPointsTransaction?> GetTransactionByReceiptIdAsync(int receiptId);
+        public Task<CreditPointsTransaction?> GetTransactionByReceiptIdAsync(BigInteger receiptId);
 
         public Task<PagedTransactionsResponse> GetTransactionsByCustomerAndRestaurantAsync(int? customerId,
             int? restaurantId, int pageNumber = 1, int pageSize = 10);
